% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-parser.r
\name{sdc.read.jv.lp}
\alias{sdc.read.jv.lp}
\title{Read Joint Ventures - List of Participants plain text reports from SDC Platinum database}
\usage{
sdc.read.jv.lp(...)
}
\arguments{
\item{sdc.file.name}{Filename of SDC Platinum plain text report}

\item{sdc.dir}{Location of SDC Platinum plain text report, sdc.dir = "../sdc-vjs" is assumed by default.}
}
\value{
A data.table where some of the columns are lists of vectors, lists and data.tables. Each row represents separate SDC record.
}
\description{
A wrapper for sdc.read(..., data.type = "jv", report.type = "lp"). It is a convinience function that makes it easear to parse and clean the data from SDC Platinum plain text reports. Several types of reports are supported. However, not everything is yet covered so suggestions, feature requests and issues reports are wellcome.
}
