% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-parser.r
\name{sdc.read}
\alias{sdc.read}
\title{Read plain text reports from SDC Platinum database}
\usage{
sdc.read(sdc.file.name, sdc.dir = getwd(), data.type = c("jv",
  "Joint Ventures"), report.type = c("csr",
  "Comprehensive Summary Report", "lp", "List of Participants"),
  fields = NA)
}
\arguments{
\item{sdc.file.name}{Filename of SDC Platinum plain text report}

\item{sdc.dir}{Location of SDC Platinum plain text report. Current working directory is assumed by default.}

\item{data.type}{Type of SDC Platinum database. Only Joint Ventures - "jv" is currently supported. M&A, Venture Experts and so forth could be added in future. Default is "jv"}

\item{report.type}{Type of plain text report. Two types are currently supported. "csr" - Comprehensive Summary Report and "lp" - List of Participants. Default is "csr".}

\item{fields}{Which fields to parse from the report. Depends on report type. For the "csr" (Comprehensive Summary Report) following fields are available - "name", "participants", "participants.details", "financial", "deal.type", "date.announced", "date.agreement", "involving", "location", "synopsis", "activity". Default value is NA which means include all fields available. Specifying just a few fields increases speed of parsing.}
}
\value{
A data.table where some of the columns are lists of vectors, lists and data.tables. Each row represents separate SDC record.
}
\description{
It is a convenience function that makes it easier to parse and clean the data from SDC Platinum plain text reports. Several types of reports are supported. However, not everything is yet covered so suggestions, feature requests and issues reports are welcome.
}
